requirements.txt
# Core dependencies
python-telegram-bot==20.7
anthropic==0.39.0
python-dotenv==1.0.0

# Tool dependencies
requests==2.31.0
google-search-results==2.4.2  # SerpAPI

# Google API dependencies (for Gmail and Calendar)
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.111.0
python-dateutil==2.8.2

# Utilities
aiohttp==3.9.1
pydantic==2.5.0